requirements.txt
pydub
ffprobe
argparse
openai>=1.0.0
python-dotenv
boto3
replicate
requests
colorama